feat: cache vendor items separately until they update
The cache is considered valid until the earliest `nextRefreshDate` returned by a vendor that we found armor items from.

This cache is currently implemented by writing to the inventory table, which is now selectively cleared when updating inventory and vendor items.
This seems to add a ~500ms delay to the inventory update, but avoids having to re-fetch all 3 vendor datasets every time.