fix: a possible fix for players reporting ghost items
The theory here would be that the reporting users are all users of an old version which was not tracking any source information,
meaning the purge of old items wasn’t catching these due to having an empty source value (but was fixed permanently by re-logging).

We can just delete everything that’s not a Vendor source here instead, which may also be faster.